# api/openapi.yaml
components:
  schemas:
    ActiveMethodView:
      description: 'One active authentication method instance. `id` addresses it for DELETE .../methods/{id} - method name
        alone isn''t unique when a method allows multiple instances (e.g. several active `device` entries, one per physical
        device). `label` is a user-chosen display name, set only for multi-instance methods; `null` for singleton ones (email/sms/password),
        which the client labels from `method` itself. `enrolledUnderAcr`/`maxAcr`/`effectiveAcr`/`factorTypes` surface the
        ADR-5 three-way cap (docs/12-entscheidungen.md): `effectiveAcr` is `min(enrolledUnderAcr, maxAcr)`, the level this
        method can actually contribute right now, which can be lower than the tool''s own declared `maxAcr` if it was enrolled
        while the session had proven less.'
      properties:
        effectiveAcr:
          description: min(enrolledUnderAcr, maxAcr) - what this method actually contributes today.
          example: loa1
          type: string
        enrolledUnderAcr:
          description: The level the session had already proven at the moment this method was enrolled (ADR-5) - caps effectiveAcr
            below maxAcr if lower.
          example: loa1
          type: string
        factorTypes:
          example:
          - POSSESSION
          items:
            enum:
            - KNOWLEDGE
            - POSSESSION
            - INHERENCE
            type: string
          type: array
        id:
          example: 7f3e2b1a-0c9d-4e8f-8a1b-2c3d4e5f6a7b
          type: string
        label:
          example: Laptop
          type: string
        maxAcr:
          description: This tool's own declared ceiling - never account/session-specific.
          example: loa2
          type: string
        method:
          example: sms
          type: string
      required:
      - id
      - method
      type: object
    AmrEntry:
      description: One native authenticator proof - which authenticator TYPE, and which specific execution/instance of it.
      properties:
        amrSourceId:
          example: kc-otp-form-exec-1
          type: string
        nativeToolId:
          example: kc-otp-form
          type: string
      required:
      - amrSourceId
      - nativeToolId
      type: object
...
    AuthData:
      properties:
        accountId:
          example: 42
          format: int64
          type: integer
        acr:
          example: loa2
          type: string
        amr:
          additionalProperties:
            type: string
          description: 'Method -> who proved it: "orchestrator" for a completed orchestrator tool, "kc" for evidence a native
            Keycloak authenticator already established (docs/05-api.md Abschnitt 3). Informational only - the orchestrator
            alone still resolves the combined acr above, regardless of source.'
          example:
            password: kc
            sms: orchestrator
          type: object
      type: object
...
    ChannelBlock:
      properties:
        activeMethods:
          description: All active authentication methods on the account, regardless of whether this session's currentAmr proved
            them. currentAmr is session evidence (what THIS channel actually proved); activeMethods is the account's full
            standing method list, unfiltered by device - a lost/stolen device's credential must be removable from any authenticated
            session, not only from that device itself.
          items:
            $ref: '#/components/schemas/ActiveMethodView'
          type: array
        channelSessionId:
          format: uuid
          type: string
        channelType:
          description: Which facade this channel was opened through - APP (DPoP) or KEYCLOAK (docs/02-domaenenmodell.md Abschnitt
            1). Fixed for the channel's whole lifetime.
          example: APP
          type: string
        currentAcr:
          example: loa2
          type: string
        currentAmr:
          example:
          - sms
          - password
          items:
            type: string
          type: array
        state:
          example: AUTHENTICATED
          type: string
      required:
      - channelSessionId
      - channelType
      - state
      type: object
...
    ChannelResponse:
      properties:
        authData:
          $ref: '#/components/schemas/AuthData'
          description: KEYCLOAK channels only (docs/05-api.md Abschnitt 3) - never present for APP.
        channel:
          $ref: '#/components/schemas/ChannelBlock'
        demo:
          $ref: '#/components/schemas/DemoInfo'
          description: Demo-only correlation IDs, never part of the production contract.
        next:
          $ref: '#/components/schemas/Next'
        stepData:
          additionalProperties:
            type: object
          description: 'Whatever the current step needs to render: missing fields, selection options, or a retry reason.'
          example:
            options:
            - ident-fsc
            - ident-eid
          type: object
      required:
      - channel
      type: object
...
    DemoInfo:
      additionalProperties: true
      properties:
        accountId:
          example: 42
          format: int64
          type: integer
        journeys:
          description: The running journey chain for this channel, outermost first - see [JourneyDebugStep]. Empty once nothing
            is running.
          items:
            $ref: '#/components/schemas/JourneyDebugStep'
          type: array
        personId:
          example: 7
          format: int64
          type: integer
      type: object
...
    JourneyDebugStep:
      properties:
        intent:
          example: DELETE_ACCOUNT
          type: string
        journeyId:
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          type: string
        lifecycle:
          example: SUSPENDED
          type: string
        note:
          description: 'Demo-only: why this journey''s current step looks the way it does - either why its tool became the
            automatic choice, or why a selection among several is being shown at all. Null whenever the step already explains
            itself (e.g. a Prompt), never part of the production contract.'
          example: Gerät wiedererkannt - automatisch vorgeschlagen.
          type: string
        stateType:
          example: ConfirmPending
          type: string
      required:
      - intent
      - journeyId
      - lifecycle
      - stateType
      type: object
...
    KcChannelUpsertRequest:
      description: Upsert body for the kc-facade's one facade-specific endpoint (docs/05-api.md Abschnitt 3). All fields are
        optional. accountId is the account Keycloak already knows (sub vorhanden) - the channel is bound to it immediately,
        once, never overwritten by a later call. targetAcr is Keycloak's requested LoA level, already translated into an orchestrator
        ACR string, and only raises the channel's floor, never lowers it. amr lists which native Keycloak authenticators (never
        orchestrator tools) just proved something THIS flow run, one entry per proof - method/loa/factorTypes are resolved
        server-side from a NativeAuthenticatorDescriptor (see AmrEntry), the kc-facade's own mirror of a ToolDescriptor, not
        resolved from the orchestrator's own catalog (which stays entirely ignorant of native authenticators). Merged into
        the channel's evidence and re-checked against the current floor exactly like any other proof; no separate 'combined
        native acr' field exists, since the orchestrator derives that itself.
      properties:
        accountId:
          example: 42
          format: int64
          type: integer
        amr:
          items:
            $ref: '#/components/schemas/AmrEntry'
          type: array
        availableTools:
          description: The Web channel's own declaration of which toolIds its Keycloak theme can render (one com.example.dpop.kcext.webtool.WebToolRenderer
            factory per toolId, registered via META-INF/services) - the kc-facade's counterpart to the App channel's own availableTools
            (POST /channels). Only read on this channel's first call (a later upsert resumes the already-persisted set); a
            channel-anonymous caller that omits this gets none of the orchestrator's tools, never all of them.
          items:
            type: string
          type: array
        intent:
          description: 'Only read on this channel''s first call, same restriction as availableTools - the kc facade''s own,
            deliberately narrow counterpart to the App facade''s `intent` request parameter (docs/05-api.md #"POST /app/channels:
            intent-Parameter"). Omitted (or null) means kc_select_method, the existing login/step-up behaviour. Only kc_select_method
            and register are accepted here - unlike the App facade, not every AuthIntent.isEntryIntent value: fast_access/lookup_login
            assume an APP-shaped channel this facade never has.'
          example: register
          type: string
        kcSessionId:
          description: Required whenever restoreData is present, ignored otherwise. Keycloak's own, durable UserSessionModel
            id - deliberately NOT read off the peer-auth assertion (the assertion's kc-anchor is always THIS flow run's own
            channelSessionId, docs/02-domaenenmodell.md Abschnitt 1, so it can't verify a token minted for a DIFFERENT, earlier
            flow run's channel). Must match what GET .../restore-data was called with to produce this exact restoreData token.
          type: string
        restoreData:
          description: 'A signed RestoreData token this same UserSession''s channel returned earlier via GET .../restore-data,
            resubmitted verbatim (docs/ideen/web-keycloak-kanal.md #6) - the bulk, one-shot way to seed a brand-new channel
            with what a PRIOR, unrelated flow run already established, as opposed to accountId/amr above which report what
            THIS flow run just proved. Both are merged into the channel the same way; only restoreData may already be meaningfully
            old by the time it arrives here. Opaque to every caller but the orchestrator itself - see RestoreDataCodec.'
          type: string
        targetAcr:
          example: loa2
          type: string
      type: object
...
    Next:
      properties:
        context:
          example: authentication
          type: string
        step:
          example: auth
          type: string
        toolId:
          example: auth-sms
          type: string
        toolSessionId:
          format: uuid
          type: string
        type:
          example: tool
          type: string
      required:
      - step
      - type
      type: object
...
    RestoreDataResponse:
      description: The channel's current RestoreData, signed - null if there is nothing worth restoring yet.
      properties:
        restoreData:
          type: string
      type: object
...
  securitySchemes:
    dpop:
      description: DPoP proof JWT (docs/09-dpop.md); required on every App-facade call.
      in: header
      name: DPoP
      type: apiKey
info:
  description: App-facing (orchestrator-first) API for the DPoP-bound registration and login demo. See docs/05-api.md for
    the full contract; the Keycloak-facing kc facade is out of scope for this build (docs/08-projektrahmen.md).
  title: DPoP-Demo Orchestrator API
  version: v1
openapi: 3.1.0
paths:
...
  /orchestrator/api/v1/channels/{channelSessionId}:
    delete:
      description: 'Ends this channel for good (docs/02-domaenenmodell.md #3: AUTHENTICATED -> LOGGED_OUT, terminal) - cancels
        any active process and discards the AuthContext. For interactive clients prefer POST .../logouts which shows a confirmation
        prompt first. Never resumes on this channelSessionId afterwards; call POST .../channels again for a new session (a
        known device still skips straight to LOGIN there).'
      operationId: logout
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      - in: query
        name: bindingKeyRef
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Logged out - no body.
      security:
      - dpop: [
          ]
      summary: Log out (direct, no confirmation)
      tags:
      - Channels
    get:
      description: 'The guaranteed resume entry point (docs/05-api.md #2): next always reflects the currently due step.'
      operationId: getChannel
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      - in: query
        name: bindingKeyRef
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  currentAcr: loa1
                  state: STEP_UP_IN_PROGRESS
                next:
                  step: auth
                  toolId: auth-sms
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Resumed mid-step-up, waiting on an SMS TAN.
      security:
      - dpop: [
          ]
      summary: Read the current channel state
      tags:
      - Channels
...
  /orchestrator/api/v1/kc/channels/{channelSessionId}:
    patch:
      description: 'Upsert semantics (docs/05-api.md Abschnitt 3): creates the channel on first call under this Keycloak-chosen
        id, just resumes it on every later one - idempotent by construction, no separate create-then-resume round trip. accountId/targetAcr
        are only meaningful on step-up.'
      operationId: upsertChannel
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      - in: header
        name: Authorization
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KcChannelUpsertRequest'
      responses:
        '200':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  state: REGISTERING
                next:
                  context: auth
                  step: selectMethod
                  type: orchestrator
                stepData:
                  options:
                  - ident-fsc
                  - auth-password
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: First call for a fresh Keycloak flow run - offers every kc-usable method.
      security:
      - kc-peer-auth: [
          ]
      summary: Upsert a kc channel and advance its journey
      tags:
      - KC channels
  /orchestrator/api/v1/kc/channels/{channelSessionId}/restore-data:
    get:
      description: For the Authenticator's end-of-flow lifecycle hook only (docs/ideen/docs/05-api.md Abschnitt 3) - reads
        back what this channel accumulated, to stash in a Keycloak UserSessionModel note and resubmit at a later flow's start.
        kcSessionId is passed explicitly (not read off the channel's own anchor) because the returned token must remain valid
        across the flow-run boundary the channel itself does not survive - a fresh channel next flow run, but the very same
        UserSessionModel id.
      operationId: getRestoreData
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      - in: query
        name: kcSessionId
        required: true
        schema:
          type: string
      - in: header
        name: Authorization
        required: false
        schema:
          type: string
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestoreDataResponse'
          description: OK
      security:
      - kc-peer-auth: [
          ]
      summary: Fetch this channel's signed RestoreData
      tags:
      - KC channels
...
  /orchestrator/api/v1/tools/{toolSessionId}/{toolId}:
    delete:
      description: Moves the journey on according to the state it is standing on - to the next fallback option, back to the
        selection step, or to the end of the journey if nothing else could be offered.
      operationId: switchAway
      parameters:
      - in: path
        name: toolSessionId
        required: true
        schema:
          format: uuid
          type: string
      - in: path
        name: toolId
        required: true
        schema:
          type: string
      - in: query
        name: bindingKeyRef
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  currentAcr: loa1
                  state: STEP_UP_IN_PROGRESS
                next:
                  context: auth
                  step: selectMethod
                  type: orchestrator
                stepData:
                  options:
                  - auth-password
                  - auth-device
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: auth-sms abandoned during a fallback chain - offers the other loa2 candidates.
      security:
      - dpop: [
          ]
      summary: Abandon this tool attempt
      tags:
      - Tools
...