# src/main/kotlin/com/example/dpop/orchestrator/api/v1/kc/KcChannelDtos.kt
package com.example.dpop.orchestrator.api.v1.kc

import com.example.dpop.orchestrator.policy.AuthEvidence
import com.fasterxml.jackson.annotation.JsonInclude
import io.swagger.v3.oas.annotations.media.Schema
import com.example.dpop.orchestrator.kernel.AuthIntent

@Schema(
    description = "Upsert body for the kc-facade's one facade-specific endpoint (docs/05-api.md " +
        "Abschnitt 3). All fields are optional. accountId is the account " +
        "Keycloak already knows (sub vorhanden) - the channel is bound to it immediately, once, " +
        "never overwritten by a later call. targetAcr is Keycloak's requested LoA level, already " +
        "translated into an orchestrator ACR string, and only raises the channel's floor, never " +
        "lowers it. amr lists which native Keycloak authenticators (never orchestrator tools) " +
        "just proved something THIS flow run, one entry per proof - method/loa/factorTypes are " +
        "resolved server-side from a NativeAuthenticatorDescriptor (see AmrEntry), the kc-facade's " +
        "own mirror of a ToolDescriptor, not resolved from the orchestrator's own catalog (which " +
        "stays entirely ignorant of native authenticators). Merged into the channel's evidence " +
        "and re-checked against the current floor exactly like any other proof; no separate " +
        "'combined native acr' field exists, since the orchestrator derives that itself."
)
data class KcChannelUpsertRequest(
    @field:Schema(example = "42")
    val accountId: Long? = null,
    @field:Schema(example = "loa2")
    val targetAcr: String? = null,
    val amr: List<AmrEntry>? = null,
    @field:Schema(
        description = "A signed RestoreData token this same UserSession's channel returned " +
            "earlier via GET .../restore-data, resubmitted verbatim (docs/ideen/web-keycloak-" +
            "kanal.md #6) - the bulk, one-shot way to seed a brand-new channel with what a PRIOR, " +
            "unrelated flow run already established, as opposed to accountId/amr above which " +
            "report what THIS flow run just proved. Both are merged into the channel the same " +
            "way; only restoreData may already be meaningfully old by the time it arrives here. " +
            "Opaque to every caller but the orchestrator itself - see RestoreDataCodec."
    )
    val restoreData: String? = null,
    @field:Schema(
        description = "Required whenever restoreData is present, ignored otherwise. Keycloak's " +
            "own, durable UserSessionModel id - deliberately NOT read off the peer-auth assertion " +
            "(the assertion's kc-anchor is always THIS flow run's own channelSessionId, " +
            "docs/02-domaenenmodell.md Abschnitt 1, so it can't verify a token minted for a DIFFERENT, " +
            "earlier flow run's channel). Must match what GET .../restore-data was called with to " +
            "produce this exact restoreData token."
    )
    val kcSessionId: String? = null,
    @field:Schema(
        description = "The Web channel's own declaration of which toolIds its Keycloak theme can " +
            "render (one com.example.dpop.kcext.webtool.WebToolRenderer factory " +
            "per toolId, registered via META-INF/services) - the kc-facade's counterpart to the App " +
            "channel's own availableTools (POST /channels). Only read on this channel's first call " +
            "(a later upsert resumes the already-persisted set); a channel-anonymous caller that " +
            "omits this gets none of the orchestrator's tools, never all of them."
    )
    val availableTools: List<String>? = null,
    @field:Schema(
        description = "Only read on this channel's first call, same restriction as availableTools " +
            "- the kc facade's own, deliberately narrow counterpart to the App facade's `intent` " +
            "request parameter (docs/05-api.md #\"POST /app/channels: intent-Parameter\"). Omitted " +
            "(or null) means kc_select_method, the existing login/step-up behaviour. Only " +
            "kc_select_method and register are accepted here - unlike the App facade, not every " +
            "AuthIntent.isEntryIntent value: fast_access/lookup_login assume an APP-shaped channel " +
            "this facade never has.",
        example = "register"
    )
    val intent: String? = null
)

/**
 * One method a native Keycloak authenticator just proved THIS flow run (docs/ideen/web-keycloak-
 * kanal.md #6/#8/#9). Deliberately just two ids, not method/loa/factorTypes too - those are fixed
 * per authenticator TYPE, so they come from [NativeAuthenticatorDescriptor] (looked up by
 * [nativeToolId]) instead of being resent on every proof, exactly like an orchestrator tool's own
 * evidence is priced from its `ToolDescriptor` rather than resent on every outcome. [nativeToolId]
 * is Keycloak's own stable authenticator/execution-CONFIG id - fixed for that authenticator's
 * whole configuration. [amrSourceId] is the stable id of THIS specific proof/execution instance
 * (its "AmrUtils"-style bookkeeping of which native step established which factor, and for how
 * long it stays valid) - carried on every `amr` report for the SAME proof, including refreshes,
 * so `AuthEvidence.AmrRecord` can remember it and a later `UpdateAuthenticator` call for the same
 * execution is recognizable as a refresh, not a new proof.
 */
@Schema(description = "One native authenticator proof - which authenticator TYPE, and which specific execution/instance of it.")
data class AmrEntry(
    @field:Schema(example = "kc-otp-form") val nativeToolId: String,
    @field:Schema(example = "kc-otp-form-exec-1") val amrSourceId: String
)

/**
 * Everything the kc-facade's `OrchestratorAuthenticator` needs to hand a brand-new channel a
 * running start (docs/05-api.md Abschnitt 3) - not just evidence, deliberately general:
 * whatever this channel accumulated that a LATER, unrelated `ChannelSession` row (a fresh flow
 * run, e.g. a step-up) should be able to resume from without re-proving it. [evidence] is the
 * REAL `AuthEvidence` this channel accumulated, not a hand-rolled duplicate of it: restoring
 * through a lossy shape of its own would drift out of sync with whatever `AuthEvidence` grows
 * next, so a restored channel's evidence is exactly what the original had.
 *
 * Fetched via its own explicit endpoint (`GET .../kc/channels/{channelSessionId}/restore-data`),
 * never bundled into the regular upsert response - the Authenticator's end-of-flow lifecycle hook
 * is the one and only caller, right when a `UserSessionModel` first becomes available to stash it
 * in a session note; every other response on the way there would carry a value nobody reads yet.
 * Kc-facade-only, same as everything else in this file - nothing outside `kc` needs it.
 *
 * Never travels on the wire as this plain shape - see [RestoreDataCodec], which signs it into an
 * opaque token bound to the UserSession it was minted for, so a leaked or reused note can't hand a
 * DIFFERENT UserSession evidence nobody actually proved under it.
 */
data class RestoreData(
    val accountId: Long? = null,
    val evidence: AuthEvidence? = null
)

/** Wire wrapper for `GET .../restore-data` (docs/05-api.md Abschnitt 3) - a plain string response body would be an unusual shape next to the rest of this JSON API. */
@JsonInclude(JsonInclude.Include.NON_NULL)
@Schema(description = "The channel's current RestoreData, signed - null if there is nothing worth restoring yet.")
data class RestoreDataResponse(val restoreData: String? = null)